+++ packages/deploy/index.ts
@@ -1,2 +1,4 @@
-export { deployNocturne, NocturneDeployOpts } from "./src/deploy";
+export { NocturneDeployConfig, NocturneDeployOpts } from "./src/config";
+export { deployNocturne } from "./src/deploy";
+export { whitelistProtocols } from "./src/whitelist";
 export { checkNocturneContractDeployment } from "./src/checks";