+++ packages/deploy/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+- separate ownership transferral from contract deployment
+- add `whitelistProtocols` function
+- remove empty `console.log()` in place of `\n` again
+- separate configs types out into separate `config.ts` again
+- add `protocolAllowlist` to example config
 - take data out of env vars and put into config.json
 - make contract owners same as proxy admin owner, in practice proxy admin owner will be gnosis safe and we want that same contract to have ownership of contracts
 - Pass weth to deposit manager initialization