+++ packages/e2e-tests/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- deploy contract with dummy info method relinquishes ownership and is used in depositAndDeploy script
+- deploy and whitelist methods in deployment use `whitelistProtocols` fn instead of custom calling handler
 - add e2e test for under-gassed tx that ensures bundler marked `OPERATION_EXECUTION_FAILED`
 - `testE2E` checks bundler op status
 - make & write localhost config to a file in `deployAndDepositScript`